From: "John Hurst" <jbhurst@ibm.net> Subject: tags with links for LinkExtor To: <libwww@perl.org> Date: Thu, 11 Nov 1999 09:31:06 +1300 Reply-To: "John Hurst" <jbhurst@ibm.net>
Hi,

I guess that the elements of the LINK_ELEMENT table in HTML::LinkExtor were
based on a W3C standard, with some extras thrown in for popular (Netscape)
extensions. According to the Quick Reference from O'Reilly's "HTML: The
Definitive Guide" 2nd Ed, the HTML tag/attributes below can have URLs (or
"links"). The ones prefixed '#' already are in LINK_ELEMENT. Is there a
policy/reason such that some of these should not be in LINK_ELEMENT?

My document had some background attributes that weren't being extracted. I
added the necessary elements to my own copy, but was wondering if you'd do
the same for the master.

#<a href=url>
<applet archive=url>
#<applet codebase=url>
#<area href=url>
#<base href=url>
<bgsound src=url>
#<body background=url>
<embed pluginspage=url>
#<embed src=url>
#<form action=url>
#<frame src=url> (twice?)
<iframe src=url>
<ilayer background=url>
<img dynsrc=url>
#<img lowsrc=url>
#<img src=url>
#<img usemap=url>
#<input [with type=image] src=url>
<isindex action=url>
<layer background=url>
<layer src=url>
#<link href=url>
<object classid=url>
<object codebase=url>
<object data=url>
<object usemap=url>
<script src=url>
<table background=url>
<td background=url>
<th background=url>
<xmp href=url>

Thanks for a great tool,

John Hurst