Test patch from Michael:
t/ignore.t was over elaborate, so I simplified it and added a test for
(handler=>undef).

I was not sure $p->parse_file would  end without an extra text event if it
was terminated by a callback invoking $p->eof, so I added tests for it in
t/handler-eof.t and t/parsefile.t.