From: Norbert Kiesel <nkiesel@tbdnetworks.com> Subject: Re: HTML::Parser: how can I reset report_tags to report all tags? To: Gisle Aas <gisle@ActiveState.com> Cc: libwww@perl.org Date: Sun, 19 Jun 2005 16:36:38 -0700 Organization: TBD Networks
Hello,

here is a patch which adds some mpre documentation for repot_tags and
ignore_tags.  I also added some test cases.

I changed the t/filter-methods.t to use Test::More, as this gives nice
error messages if tests fail (which they did while I was trying to
figure this stuff out :-).  Only drawback is that this makes it
dependent on Test::More, but I think that is used by many other modules
as well.  If you think that's unacceptable, I'd be willing to rewrite it
without using Test::More.

Best,
  Norbert