From: Norbert Kiesel <nkiesel@tbdnetworks.com> Subject: Re: HTML::Parser: how can I reset report_tags to report all tags? To: Gisle Aas <gisle@ActiveState.com> Date: Tue, 21 Jun 2005 11:57:27 -0700 Organization: TBD Networks
On Tue, 2005-06-21 at 00:19 -0700, Gisle Aas wrote:
> I've usually stayed with Test.pm to be as portable as possible, but
> since I already applied one Test::More patch more of those should not
> hurt.  Please send them along.

You asked for it, so don't blame me for a big patch :-)

Here is what I did:
- changed all tests to use Test::More
- changed all debug "print" to "diag"
- dropped trailing "\n" from diag (done by diag)
- split up some && tests into separate tests

I tested this under my Debian unstable (which has a very recent perl
environment) as well on an older Sun box (but also running perl 5.8.3),
and all tests succeed. "make test 2>/dev/null" now only prints the test
results (as "diag" prints to stderr).

As Test::More::is() already prints the expected and the real value if
the test fails, some of the "diag" statements could probably now be
removed, but this is not done yet.  Another improvement could be to name
these tests (as I did for filter-methods.t).

Best,
  Norbert